requirements.txt

pip>=23.2.1
openai
Flask==2.1.2
Werkzeug==2.1.0
torch==1.10.2
torchvision==0.11.3
gunicorn
Pillow==9.0.1

requests
python-dotenv
numpy